[JDZH3-94/96] bug fixes

+++ src/main/java/com/cydeo/controller/ClientVendorController.java
@@ -1,12 +1,11 @@
 package com.cydeo.controller;
 
 import com.cydeo.dto.ClientVendorDTO;
-import com.cydeo.entity.ClientVendor;
 import com.cydeo.enums.ClientVendorType;
 import com.cydeo.mapper.MapperUtil;
+import com.cydeo.repository.InvoiceRepository;
 import com.cydeo.service.ClientVendorService;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.catalina.LifecycleState;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -22,10 +21,13 @@ import java.util.List;
 public class ClientVendorController {
 
     private final ClientVendorService clientVendorService;
+
+    private final InvoiceRepository invoiceRepository;
     private final MapperUtil mapperUtil;
 
-    public ClientVendorController(ClientVendorService clientVendorService, MapperUtil mapperUtil) {
+    public ClientVendorController(ClientVendorService clientVendorService, InvoiceRepository invoiceRepository, MapperUtil mapperUtil) {
         this.clientVendorService = clientVendorService;
+        this.invoiceRepository = invoiceRepository;
         this.mapperUtil = mapperUtil;
     }
 
@@ -76,14 +78,20 @@ public class ClientVendorController {
             model.addAttribute("clientVendor", clientVendorService.findById(clientVendorId));
             model.addAttribute("clientVendorTypes", Arrays.asList(ClientVendorType.values()));
             return "clientVendor/clientVendor-update";
-
         }
-        clientVendorService.update(clientVendorDTO, clientVendorId);
+        clientVendorDTO.setId(clientVendorId);
+        clientVendorService.update(clientVendorDTO);
         return "redirect:/clientVendors/list";
 
     }
     @GetMapping("/delete/{clientVendorId}")
-    public String deleteClientVendorById(@PathVariable("clientVendorId") Long clientVendorId, Model model){
+    public String deleteClientVendorById(@PathVariable("clientVendorId") Long clientVendorId, Model model) throws IllegalAccessException {
+        if (invoiceRepository.existsById(clientVendorId)){
+//            String error="cannot delete client/vendor linked to open invoice"; how to return this in HTML page
+//            model.addAttribute("error", error);
+            return "redirect:/clientVendors/list";
+        }
+     //trying to add above
         model.addAttribute("clientVendor", clientVendorService.findById(clientVendorId));
         clientVendorService.delete(clientVendorId);
         return "redirect:/clientVendors/list";
